excel maximum row

+++ src/py/main.py
@@ -12,14 +12,13 @@ from constant import svc_delimitor, file_config_name, existing_dell_asset_dir, s
 from excel import save_dell_asset_excel
 import sys, traceback
 
-required_arg_list = ['--parent_path=', '--svctag=', '--job_mode=', '--v=']
+required_arg_list = ['--parent_path=', '--svctag=', '--v=']
 subject_temp = "%s 时间%s 标签%s"
 
 if __name__ == "__main__":
 	arguments = parse_cmd_args(sys.argv, required_arg_list)
 	parent_path = arguments['parent_path']
 	svctag = arguments['svctag']
-	job_mode = arguments['job_mode']
 	verbose = True if 'v' in arguments else False
 	logger = Logger(verbose)
 	start_time = get_current_datetime()
@@ -34,7 +33,7 @@ if __name__ == "__main__":
 		history_valid_svctag_path = parent_path + "valid_svctags.txt"
 		dell_asset_path = existing_dell_asset_dir
 		search_history_path = parent_path + "search_history.yml"
-		dell_asset_output_path = parent_path + "output_%s.xls" % svctag
+		dell_asset_output_path = parent_path + "output/output_%s.xls" % svctag
 		api_url = config['dell_api_url']
 		api_key_L = config["dell_api_key"].values()
 		transl_url = config[config_translation_url]
@@ -45,52 +44,47 @@ if __name__ == "__main__":
 		additional_text = ""
 		need_translation = False
 		try:
-			if job_mode == job_mode_dell_asset:
-				subject = subject_temp % ('新的查询开始', start_time, svctag)
-				send_email(subject=subject, text="===请等待邮件结果", config=config)
-				logger.info("Send email to %s" % config['mail_to'])	
-				target_svc_L, existing_svc_S = target_svctags_batch(svc_L, dell_support_url, dell_asset_path, history_valid_svctag_path, logger, search_history_path)
-				# Use valid service tags to call Dell API, and parse JSON data into a list of DellAsset entities
-				if len(target_svc_L) == 0:
-					logger.warn("No target svctag needed for API call")
+			subject = subject_temp % ('新的查询开始', start_time, svctag)
+			send_email(subject=subject, text="===请等待邮件结果", config=config)
+			logger.info("Send email to %s" % config['mail_to'])	
+			target_svc_L, existing_svc_S = target_svctags_batch(svc_L, dell_support_url, dell_asset_path, history_valid_svctag_path, logger, search_history_path)
+			# Use valid service tags to call Dell API, and parse JSON data into a list of DellAsset entities
+			if len(target_svc_L) == 0:
+				logger.warn("No target svctag needed for API call")
+			else:
+				api_dell_asset_L = api_entities_batch(target_svc_L, api_url, api_key_L, logger)
+				if len(api_dell_asset_L) > 0:
+					output_dell_asset_L, NA_dict = translate_dell_warranty(transl_url, api_dell_asset_L, logger)
 				else:
-					api_dell_asset_L = api_entities_batch(target_svc_L, api_url, api_key_L, logger)
-					if len(api_dell_asset_L) > 0:
-						output_dell_asset_L, NA_dict = translate_dell_warranty(transl_url, api_dell_asset_L, logger)
-					else:
-						logger.warn("======No data for Dell Asset from API call")
-				if len(existing_svc_S) > 0:
-					existing_dell_asset_L = DellAsset.parse_dell_asset_file_batch(dell_asset_path, existing_svc_S, logger=logger)
-					# Translate all Warranties of each DellAsset, and find those warranties without available translation
-					updated_dell_asset_L, NA_dict2 = update_dell_warranty_translation(transl_url, existing_dell_asset_L, dell_asset_path, logger)
-					output_dell_asset_L.extend(updated_dell_asset_L)
-					NA_dict.update(NA_dict2)
+					logger.warn("======No data for Dell Asset from API call")
+			if len(existing_svc_S) > 0:
+				existing_dell_asset_L = DellAsset.parse_dell_asset_file_batch(dell_asset_path, existing_svc_S, logger=logger)
+				# Translate all Warranties of each DellAsset, and find those warranties without available translation
+				updated_dell_asset_L, NA_dict2 = update_dell_warranty_translation(transl_url, existing_dell_asset_L, dell_asset_path, logger)
+				output_dell_asset_L.extend(updated_dell_asset_L)
+				NA_dict.update(NA_dict2)
+			else:
+				logger.warn("======No existing Dell Asset for service tag " + svctag)
+			if verify_NA_translation(NA_dict, logger):
+				logger.warn("查询结果存在保修需要翻译")
+				need_translation = True
+				additional_text += "查询结果存在保修需要翻译"
+			else:
+				logger.info("No additional translation needed")
+			if len(output_dell_asset_L) > 0:
+				logger.info("~~~~~~~%s output results in total" % len(output_dell_asset_L))
+				# Save output into the csv_path and also existing dell asset
+				logger.info("~~~~~~~Save output as existing dell assets")
+				for da in output_dell_asset_L:
+					if da is not None and da.svctag != "" and da.is_translation_updated:
+						temp_path = existing_dell_asset_dir + da.svctag + history_DA_file_format
+						save_object_to_path(value=da, output_path=temp_path)
+				if save_dell_asset_excel(output_dell_asset_L, dell_asset_output_path):
+					logger.info("保存结果到Excel文件 %s" % dell_asset_output_path)
 				else:
-					logger.warn("======No existing Dell Asset for service tag " + svctag)
-				if verify_NA_translation(NA_dict, logger):
-					logger.warn("查询结果存在保修需要翻译")
-					need_translation = True
-					additional_text += "查询结果存在保修需要翻译"
-				else:
-					logger.info("No additional translation needed")
-				if len(output_dell_asset_L) > 0:
-					logger.info("~~~~~~~%s output results in total" % len(output_dell_asset_L))
-					# Save output into the csv_path and also existing dell asset
-					logger.info("~~~~~~~Save output as existing dell assets")
-					for da in output_dell_asset_L:
-						if da is not None and da.svctag != "" and da.is_translation_updated:
-							temp_path = existing_dell_asset_dir + da.svctag + history_DA_file_format
-							save_object_to_path(value=da, output_path=temp_path)
-					if save_dell_asset_excel(output_dell_asset_L, dell_asset_output_path):
-						logger.info("保存结果到Excel文件 %s" % dell_asset_output_path)
-					else:
-						logger.error("保存Excel文件出错")
-				else:
-					logger.warn("-------Output for this job is empty")
-			elif job_mode == job_mode_update_svctag:
-				subject = subject_temp % ('新的标签更新开始', start_time, svctag)
-				target_svctags_batch(svc_L, dell_support_url, dell_asset_path, history_valid_svctag_path, logger, svc_job=True)
-				send_email(subject=subject, text=logger, config=config)
+					logger.error("保存Excel文件出错")
+			else:
+				logger.warn("-------Output for this job is empty")
 		except Exception, e:
 			logger.error(str(e))
 			logger.error(traceback.format_exc())
@@ -108,6 +102,5 @@ if __name__ == "__main__":
 			else:
 				logger.error("Send email failed")
 			save_object_to_path(value=logger, output_path=log_output_path)
+		if not logger.has_error:
 			delete_file(dell_asset_output_path)
-		elif job_mode == job_mode_update_svctag:
-			pass
